Trying to fix type error

+++ kipoi/model.py
@@ -995,7 +995,7 @@ class PyTorchModel(BaseModel, GradientMixin, LayerActivationMixin):
         input = self.numpy_to_torch(x, requires_grad=requires_grad)
         if isinstance(x, np.ndarray):
             # convert to a pytorch tensor and then to a pytorch variable
-            pred = self.model(input)
+            pred = self.model(input.astype(np.float32))
 
         elif isinstance(x, dict):
             # convert all entries in the dict to pytorch variables
